Convert to FastAPI and kubernetes_asyncio
Convert Moneypenny to FastAPI from aiohttp.  Use Pydantic for input
validation instead of a JSON schema.  Update some boilerplate to
the standard FastAPI template.

Convert the Kubernetes layer to kubernetes_asyncio so that it won't
block other operations while Kubernetes operations are in progress.

Use FastAPI BackgroundTasks instead of aiojobs to kick off the
creation and management of the Kubernetes pod that does the work.
Use a workaround from StackOverflow to inject those tasks into the
main asyncio loop rather than running them as normal background tasks
to work around a limitation in use of httpx for testing.

Mock Kubernetes for testing, allowing the full test suite to be run
without a Kubernetes cluster.