Fix race condition on commission/retire
The previous implementation was returning success and then creating
the pod as a background task.  Unfortunately, this meant we raced
with the client polling status, and might return 404 because the
pods hadn't been created yet now that the Kubernetes code is also
async.

Fix this by changing the API to start the pod before returning a
response and returning a redirect to the status endpoint instead of
a 202 response.  Only background the code that waits for the pod to
complete and then cleans up.