Stop sharing one KubernetesClient
The documentation for kubernetes_asyncio shows each async worker
creating a separate client with async with, so follow the same
pattern in the moneypenny KubernetesClient.  This allows creating
a new KubernetesClient per request; take advantage of that by
passing in a logger with additional context information.

Move some generic Kubernetes functions out of the KubernetesClient
class for eventual move to Safir.