Use a Kubernetes client for each operation
Trying to use a Kubernetes client for each Moneypenny object causes
the client to be closed before background processing completes.
Instead, open and close clients inside KubernetesClient for each
operation and hope that this won't be too slow.  It unforuntately
means that we are recycling the connection pool, so we may want to
revisit this later.