Handle multiple requests for the same action
If someone requests commissioning while it is still in progress but
asks for the same parameters, just redirect to the status page as if
we created a new pod and let them do the normal wait for the existing
spawned pod.

Move some messages to debug that aren't important to see all the time.